fuse: allow using readdir cache
The cache is only used if it's completed, not while it's still being
filled; this constraint could be lifted later, if it turns out to be
useful.

Introduce state in struct fuse_file that indicates the position within the
cache.  After a seek, reset the position to the beginning of the cache and
search the cache for the current position.  If the current position is not
found in the cache, then fall back to uncached readdir.

It can also happen that page(s) disappear from the cache, in which case we
must also fall back to uncached readdir.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
(cherry picked from commit 5d7bc7e8680c7ca4c8a4f139ce2a54ccb8131ef0)
[keyur: 4.9 backport - the new fuse_readdir() wrapper uses fuse_is_bad(), not
is_bad_inode(): 4.9.y stable already carries the fuse bad-inode conversion that
this commit predates.]