fuse: enable caching of symlinks
FUSE file reads are cached in the page cache, but symlink reads are
not. This patch enables FUSE READLINK operations to be cached which
can improve performance of some FUSE workloads.

In particular, I'm working on a FUSE filesystem for access to source
code and discovered that about a 10% improvement to build times is
achieved with this patch (there are a lot of symlinks in the source
tree).

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
(cherry picked from commit 5571f1e65486be025f73fa6aa30fb03725d362a2)
[keyur: 4.9 backport - fuse_get_link() uses fuse_is_bad(), not is_bad_inode():
4.9.y stable already carries the fuse bad-inode conversion that this commit
predates. Downstream FUSE_PASSTHROUGH keeps bit 31.]