convenience helpers: vfs_get_super() and sget_fc()
the former is an analogue of mount_{single,nodev} for use in
->get_tree() instances, the latter - analogue of sget() for the
same.

These are fairly similar to the originals, but the callback signature
for sget_fc() is different from sget() ones, so getting bits and
pieces shared would be too convoluted; we might get around to that
later, but for now let's just remember to keep them in sync.  They
do live next to each other, and changes in either won't be hard
to spot.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
[keyur: backport to 4.9 - sget_fc() uses a local destroy_unused_super() built
 on 4.9's destroy_super() (upstream's version also frees the preallocated
 shrinker, which 4.9 has no concept of) and plain register_shrinker() rather
 than register_shrinker_prepared().  Avoids dragging in the 4.19 alloc_super()
 refactor and the mm shrinker-preallocation series, neither of which fuse
 needs.]

(cherry picked from commit cb50b348c71ffa90d7d1b2a494b553b5099bc090)