vfs: subtype handling moved to fuse
The unused vfs code can be removed.  Don't pass empty subtype (same as if
->parse callback isn't called).

The bits that are left involve determining whether it's permitted to split the
filesystem type string passed in to mount(2).  Consequently, this means that we
cannot get rid of the FS_HAS_SUBTYPE flag unless we define that a type string
with a dot in it always indicates a subtype specification.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
(cherry picked from commit c7eb6869632a5d33b41d0a00d683b8395392b7ee)
[keyur: 4.9 backport - do_new_mount() here is not fs_context-converted, so it keeps
the legacy vfs_kern_mount() + fs_set_subtype() path and can still produce an empty
s_subtype; the show_type() guard on s_subtype[0] is kept for that reason. fc->subtype
is removed from the VFS as upstream does. fs_context has no ->security member in this
tree (4.9 LSM uses ->secdata), so the fsopen(2) reconfigure helpers are not added.]