fuse: split fuse_mount off of fuse_conn
We want to allow submounts for the same fuse_conn, but with different
superblocks so that each of the submounts has its own device ID.  To do
so, we need to split all mount-specific information off of fuse_conn
into a new fuse_mount structure, so that multiple mounts can share a
single fuse_conn.

We need to take care only to perform connection-level actions once (i.e.
when the fuse_conn and thus the first fuse_mount are established, or
when the last fuse_mount and thus the fuse_conn are destroyed).  For
example, fuse_sb_destroy() must invoke fuse_send_destroy() until the
last superblock is released.

To do so, we keep track of which fuse_mount is the root mount and
perform all fuse_conn-level actions only when this fuse_mount is
involved.

Signed-off-by: Max Reitz <mreitz@redhat.com>
Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
(cherry picked from commit fcee216beb9c15c3e1466bb76575358415687c55)

[keyur: core hunks only - fs/fuse/dax.c and fs/fuse/virtio_fs.c are not backported.
The downstream code moves with the split: fuse_file carries a fuse_mount, so the CAF
passthrough (passthrough.c, fuse_send_open) and the canonical-path request go through
ff->fm / get_fuse_mount(). The async_dio short-read optimisation is NOT reintroduced:
this tree already carries "fuse: fix the ->direct_IO() treatment of iov_iter", which
replaced it with the iov_iter shorten/reexpand logic.]